新增了对token的支持，对部分接口进行保护

+++ src/main/java/com/zjgsu/online_market/entity/Good.java
@@ -2,6 +2,9 @@ package com.zjgsu.online_market.entity;
 
 import java.math.BigDecimal;
 import java.io.Serializable;
+
+import com.baomidou.mybatisplus.annotation.TableId;
+import com.baomidou.mybatisplus.annotation.TableName;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.experimental.Accessors;
@@ -27,7 +30,7 @@ public class Good implements Serializable {
 
     private String image;
 
-    private BigDecimal price;
+    private Double price;
 
     private Integer stock;
