+++ src/main/java/com/zjgsu/online_market/mapper/GoodMapper.java
@@ -19,7 +19,7 @@ import java.util.List;
 @Mapper
 public interface GoodMapper extends BaseMapper<Good>{
     @Insert("insert into good(uid,gname,description,price,stock,image,status) values(#{uid},#{gname},#{description},#{price},#{stock},#{image},#{status})")
-    void publishGood(@Param("uid") Long uid, @Param("gname") String gname, @Param("description") String description, @Param("price") BigDecimal price, @Param("stock") Integer stock, @Param("image") String image,@Param("status") Integer status);
+    void publishGood(@Param("uid") Long uid, @Param("gname") String gname, @Param("description") String description, @Param("price") Double price, @Param("stock") Integer stock, @Param("image") String image,@Param("status") Integer status);
     @Update("update good set status = #{status}, stock = #{stock} where gid = #{gid}")
     void updateGoodStatus_sell(@Param("gid") Long gid,@Param("stock") Integer stock, @Param("status") Integer status);
 }
